requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
pydantic==2.7.1
google-generativeai==0.5.2
chromadb==0.5.0
python-dotenv==1.0.1
numpy==1.26.4
